% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arcgisrouting-package.R
\docType{package}
\name{arcgisrouting-package}
\alias{arcgisrouting}
\alias{arcgisrouting-package}
\title{arcgisrouting: Access the ArcGIS Routing and Network Analysis Services}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Bindings to the 'ArcGIS' Routing REST API (\url{https://developers.arcgis.com/rest/routing/}) for solving network analysis problems. Plan routes and generate driving directions, measure travel time and distance with origin-destination cost matrices, build service areas, find the closest facilities, route fleets of vehicles, and snap GPS tracks to roads. Both synchronous requests and asynchronous geoprocessing jobs are supported, returning simple features ('sf') objects ready for analysis and mapping.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://r.esri.com/arcgisrouting/}
  \item \url{https://github.com/r-arcgis/arcgisrouting}
  \item Report bugs at \url{https://github.com/r-arcgis/arcgisrouting/issues}
}

}
\author{
\strong{Maintainer}: Josiah Parry \email{josiah.parry@gmail.com} (\href{https://orcid.org/0000-0001-9910-865X}{ORCID})

}
\keyword{internal}
