% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job-find-closest-facilities.R
\name{as_cf_locations}
\alias{as_cf_locations}
\alias{as_incidents}
\alias{as_facilities}
\title{Convert spatial objects to incidents or facilities for closest facility}
\usage{
as_incidents(x, ...)

as_facilities(x, ...)
}
\arguments{
\item{x}{An \code{sf} or \code{sfc} object containing point geometries.}

\item{...}{Additional arguments passed to methods.}

\item{verbose}{Logical. Whether to message which recognized attributes are
used. Default: \code{TRUE}.}
}
\description{
Convert spatial objects to incidents or facilities for closest facility
}
