% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-lmd-orders.R
\name{as_lmd_orders}
\alias{as_lmd_orders}
\title{Convert spatial objects to Last Mile Delivery orders input}
\usage{
as_lmd_orders(x, ...)
}
\arguments{
\item{x}{An \code{sf} or \code{sfc} object containing point geometries representing
delivery and pickup locations.}

\item{...}{Additional arguments passed to methods.}

\item{verbose}{Logical. Whether to message which recognized attributes are
used. Default: \code{TRUE}.}
}
\description{
Convert spatial objects to Last Mile Delivery orders input
}
