% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-lmd-routes.R
\name{as_lmd_routes}
\alias{as_lmd_routes}
\title{Convert a data.frame to Last Mile Delivery routes input}
\usage{
as_lmd_routes(x, ...)
}
\arguments{
\item{x}{A \code{data.frame} describing vehicle and driver characteristics,
or \code{NULL}.}

\item{...}{Additional arguments passed to methods.}
}
\description{
Convert a data.frame to Last Mile Delivery routes input
}
