% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-od-points.R
\name{as_od_points}
\alias{as_od_points}
\title{Convert spatial objects to origin-destination points}
\usage{
as_od_points(x, ...)
}
\arguments{
\item{x}{An \code{sfc} or \code{sf} object containing point geometries.}

\item{...}{Additional arguments passed to methods.}
}
\description{
Convert spatial objects to origin-destination points
}
\keyword{internal}
