% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-lmd-specialties.R
\name{as_order_specialties}
\alias{as_order_specialties}
\title{Convert a data.frame to Last Mile Delivery order specialties input}
\usage{
as_order_specialties(x, ...)
}
\arguments{
\item{x}{A \code{data.frame} with columns \code{order_name} and \code{specialty_name},
or \code{NULL}.}

\item{...}{Additional arguments passed to methods.}
}
\description{
Convert a data.frame to Last Mile Delivery order specialties input
}
