% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-polyline-barriers.R
\name{as_polyline_barriers}
\alias{as_polyline_barriers}
\title{BarrierType:
0 (restriction)
2 (added cost)
Must have accompanying attribute
Valid attributes Attr_TravelTime , Attr_Miles , Attr_Kilometers , Attr_Minutes , Attr_WalkTime , Attr_TruckMinutes , or Attr_TruckTravelTime
Convert spatial objects to polyline barriers input}
\usage{
as_polyline_barriers(x, ...)
}
\arguments{
\item{x}{An \code{sf} or \code{sfc} object containing line geometries, or \code{NULL}.}

\item{...}{Additional arguments passed to methods.}
}
\description{
BarrierType:
0 (restriction)
2 (added cost)
Must have accompanying attribute
Valid attributes Attr_TravelTime , Attr_Miles , Attr_Kilometers , Attr_Minutes , Attr_WalkTime , Attr_TruckMinutes , or Attr_TruckTravelTime
Convert spatial objects to polyline barriers input
}
\keyword{internal}
