% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-vrp-routes.R
\name{as_routes}
\alias{as_routes}
\title{Convert a data.frame to VRP routes input}
\usage{
as_routes(x, ...)
}
\arguments{
\item{x}{A \code{data.frame} describing vehicle and driver characteristics,
or \code{NULL}.}

\item{...}{Additional arguments passed to methods.}
}
\description{
Convert a data.frame to VRP routes input
}
