% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-stops.R
\name{as_stops}
\alias{as_stops}
\title{Convert spatial objects to stops for routing}
\usage{
as_stops(x, ...)
}
\arguments{
\item{x}{An \code{sfc} or \code{sf} object containing point geometries}

\item{...}{Additional arguments passed to methods}
}
\description{
Convert spatial objects to stops for routing
}
