% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-vrp-orders.R
\name{as_vrp_orders}
\alias{as_vrp_orders}
\title{Convert spatial objects to VRP orders input}
\usage{
as_vrp_orders(x, ...)
}
\arguments{
\item{x}{An \code{sfc} or \code{sf} object containing point geometries.}

\item{...}{Additional arguments passed to methods.}
}
\description{
Convert spatial objects to VRP orders input
}
