% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decode-compressed-geometry.R
\name{decode_compressed_geometry}
\alias{decode_compressed_geometry}
\title{Decode Compressed Geometry}
\usage{
decode_compressed_geometry(geometry)
}
\arguments{
\item{geometry}{Character vector containing compressed geometry strings.}
}
\value{
An \code{sfc} object with \code{LIENSTRING} geometries in EPSG:4326.
}
\description{
Decodes ArcGIS compressed geometry strings and converts them to an sfc object.
}
\references{
\href{https://github.com/maslke/arcgis-compressed-geometry}{maslke/arcgis-compressed-geometry}
}
