% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job-interface.R
\name{esri_job}
\alias{esri_job}
\title{Create a Geoprocessing Job}
\description{
Create a Geoprocessing Job

Create a Geoprocessing Job
}
\details{
The \code{gp_job} class is used to interact with Geoprocessing Services in
ArcGIS Online and Enterprise.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{base_url}}{the URL of the job service (without \verb{/submitJob})}

\item{\code{id}}{the ID of the started job. \code{NULL} \code{self$start()} has not been called.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{params}}{returns an S7 object of class \link{esri_form_params} the list can be accessed via \code{self$params@params}.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-gp_job-new}{\code{esri_job$new()}}
\item \href{#method-gp_job-start}{\code{esri_job$start()}}
\item \href{#method-gp_job-cancel}{\code{esri_job$cancel()}}
\item \href{#method-gp_job-clone}{\code{esri_job$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gp_job-new"></a>}}
\if{latex}{\out{\hypertarget{method-gp_job-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{esri_job$new(base_url, params = list(), token = arc_token())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{base_url}}{the URL of the job endpoint (without \verb{/submitJob})}

\item{\code{params}}{a named list where each element is a scalar character}

\item{\code{token}}{default \code{\link[arcgisutils:token]{arcgisutils::arc_token()}}. The token to be used with the job.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gp_job-start"></a>}}
\if{latex}{\out{\hypertarget{method-gp_job-start}{}}}
\subsection{Method \code{start()}}{
Starts the job by calling the \verb{/submitJob} endpoint. This also sets the public field \code{id}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{esri_job$start()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gp_job-cancel"></a>}}
\if{latex}{\out{\hypertarget{method-gp_job-cancel}{}}}
\subsection{Method \code{cancel()}}{
Cancels a job by calling the \verb{/cancel} endpoint.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{esri_job$cancel()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gp_job-clone"></a>}}
\if{latex}{\out{\hypertarget{method-gp_job-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{esri_job$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
