% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-travel-mode.R
\name{get_travel_modes}
\alias{get_travel_modes}
\title{Get available travel modes}
\usage{
get_travel_modes(token = arc_token())
}
\arguments{
\item{token}{Authorization token. Default: \code{\link[arcgisutils:token]{arcgisutils::arc_token()}}.}
}
\value{
A character vector of supported travel mode names.
}
\description{
Returns the names of the travel modes supported by the routing services
associated with the provided token.
}
\examples{
\dontrun{
# This example is not executed since it requires a network connection
# to ArcGIS Online and a valid authentication token
library(arcgisutils)
set_arc_token(auth_user())

get_travel_modes()
}
}
\concept{travel modes}
