% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-retrieve-travel-modes.R
\name{retrieve_travel_modes}
\alias{retrieve_travel_modes}
\title{Retrieve Available Travel Modes}
\usage{
retrieve_travel_modes(
  token = arcgisutils::arc_token(),
  error_call = rlang::caller_call()
)
}
\arguments{
\item{token}{Authorization token. Default: \code{\link[arcgisutils:token]{arcgisutils::arc_token()}}.}
}
\value{
A list describing the supported travel modes, including a
\code{supportedTravelModes} data frame.
}
\description{
Retrieves the travel modes supported by the routing services associated with
the provided token.
}
\seealso{
Other travel modes: 
\code{\link{get_travel_modes}()}
}
\concept{travel modes}
