% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arcgisrouting-package.R
\docType{package}
\name{arcgisrouting-package}
\alias{arcgisrouting}
\alias{arcgisrouting-package}
\title{arcgisrouting: What the Package Does (One Line, Title Case)}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{http://r.esri.com/arcgisrouting/}
  \item \url{https://github.com/r-arcgis/arcgisrouting}
  \item Report bugs at \url{https://github.com/r-arcgis/arcgisrouting/issues}
}

}
\author{
\strong{Maintainer}: Josiah Parry \email{josiah.parry@gmail.com} (\href{https://orcid.org/0000-0001-9910-865X}{ORCID})

}
\keyword{internal}
