% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arcgisrouting-package.R
\docType{package}
\name{arcgisrouting-package}
\alias{arcgisrouting}
\alias{arcgisrouting-package}
\title{arcgisrouting: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Josiah Parry \email{josiah.parry@gmail.com} (\href{https://orcid.org/0000-0001-9910-865X}{ORCID})

}
\keyword{internal}
